test: add parseJsonSafely tests
